+++ CHANGELOG.md
@@ -13,11 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new library X10D.Hosting.
 - Added .NET 7 target.
 - X10D: Added `IntrinsicExtensions` and `IntrinsicUtility` which offer methods for vectors in `System.Runtime.Instrinsics`. (#70)
-- X10D: Added `MathUtility.Bias(float, float)`.
+- X10D: Added `MathUtility.Bias(float, float)` and `MathUtility.Bias(double, double)`.
 - X10D: Added `MathUtility.InverseLerp(float, float, float)` and `MathUtility.InverseLerp(double, double, double)`
 - X10D: Added `MathUtility.ScaleRange(float, float, float, float, float)`
   and `MathUtility.ScaleRange(double, double, double, double, double)`
 - X10D:
+- X10D: Added `MathUtility.SmoothStep(float, float, float)` and `MathUtility.SmoothStep(double, double, double)`.
   Added `Circle`, `CircleF`, `Cuboid`, `Ellipse`, `EllipseF`, `Line3D`, `Line`, `LineF`, `Polygon`, `PolygonF`, `Polyhedron`,
   and `Sphere`, to complement System.Drawing structs such as `Point` and `Rectangle`
 - X10D: Added `Color.Deconstruct()` - with optional alpha parameter.